+++ NEWS.md
@@ -3,6 +3,8 @@
 # rquery 0.4.4 2018/05/10
 
  * Minor documentation fixes.
+ * Fix if_else NA treatment.
+ * Better detection of column types: dbi_coltypes().
 
 # rquery 0.4.3 2018/05/07